Install react and create top level component
