+++ package.json
@@ -59,6 +59,8 @@
     "jimp": "^0.2.24",
     "lodash": "^4.13.1",
     "node-memecaptain-api": "git+https://github.com/whilestevego/node-memecaptain-api.git",
+    "react": "^15.1.0",
+    "react-dom": "^15.1.0",
     "request": "^2.67.0"
   },
   "engines": {
