+++ src/renderer/components/app.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function App (props) {
+  return (
+    <section className='app'>
+      Stuff!!!!
+    </section>
+  )
+}
