+++ src/renderer/index.html
@@ -8,25 +8,7 @@
     <link href="../assets/stylesheets/base.css" rel="stylesheet" type="text/css"/>
   </head>
   <body>
-
-    <section>
-      <input
-        id="search-query"
-        class="form-control"
-        type="text"
-        placeholder="Consult an Advice Animal"
-      />
-    </section>
-
-    <section>
-      <img
-        alt="Advisor Logo"
-        id="advice-animal"
-        class="placeholder"
-        src="../assets/images/doge-icon-512.png"
-      />
-    </section>
-
-    <script type="text/javascript" src="./index.js"></script>
+      <div id="root"></div>
+      <script>require("./index.js")</script>
   </body>
 </html>
